+++ R/multisimsum.R
@@ -63,6 +63,13 @@ multisimsum <- function(data,
     checkmate::reportAssertions(collection = arg_checks)
   }
 
+  ### Coerce `methodvar` to character (if specified and not already string)
+  if (!is.null(methodvar)) {
+  	if (class(data[[methodvar]]) != "character") {
+  		data[[methodvar]] = as.character(data[[methodvar]])
+  	}
+  }
+
   ### Set reference method if `ref` is not specified
   if (!is.null(methodvar)) {
     methods <- sort(unique(data[[methodvar]]))